Added function to spit warning message
The practice column in the final output determines the individual report a record will fall into. If dataframe has missing practices, this will result in the record not filtering into the proper report causing us to miss the hospital utilization. I added a function to spit a warning message if there are missing practices. This will make this error more visible for others new to the process.